Added Orders and Products Section in Sidebar  & Signout Feature

+++ src/actions/auth.actions.js
@@ -56,11 +56,19 @@ export const isUserLoggedIn = () => {
 
 export const signout = () => {
     return async dispatch => {
-        console.log("wallah");
-        localStorage.clear();
+        dispatch({type:authConstants.LOGOUT_REQUEST});
+        const res = await axios.post('/admin/signout');
+        if(res.status === 200){
+            localStorage.clear(); 
         dispatch({
-            type:authConstants.LOGOUT_REQUEST
+            type:authConstants.LOGOUT_SUCCESS
         });
+    }else{
+        dispatch({
+            type:authConstants.LOGOUT_FAILURE,
+            payload:{error:res.data.error}
+        });
+    }
     }
 }
